Setup demo-room
Updated meshes such as SM_WallCornerIn and SM_WallCornerOut as well as the pivot point on most meshes to be easier to place. Created colliders and prefabs for Floor3x3, Wall3x3, WallCornerIn, WallCornerOut, WallDoorframe, Bureau, Couch, Player, Camera. Replaced old assets with the new prefabs. Imported scripts from Christoffer for player controller and camera controller so that player and camera now can move in demo room. Replaced skybox to just a dark color and dimmed the directional lights not to be the main light source but to still give ambient light. Added some point lights randomly in the level just to represent what it could look like if we lit with wall mounted candles later on. placed a bright pointlight on as a child to player as well, as if the player keeps a candle with them. Think that's it for now. And setup the feature branch demo_room for these changes.